Make data-grid self-contained below the visualizations module
- PaginationFooter moves into data-grid, slimmed to take a preformatted
  message prop and styled with a CSS module instead of emotion; the
  row-range and item messages are now computed by the callers
  (TableInteractive and ObjectDetail).
- The Footer no longer imports the row-count formatter; TableInteractive
  supplies formatRowsCountMessage and formatPaginationMessage props
  through DataGrid, replacing the rowsTruncated prop.
- useForceUpdate and the keyboard helpers (isPlainKey, hasModifierKeys)
  move from metabase/common down to metabase/utils; they are generic and
  data-grid was reaching up for them.

getPortalRootElement stays in metabase/common: it depends on the
embedding-sdk portal root constant, so it isn't generic enough for
metabase/utils.
