+++ frontend/src/metabase/data-grid/components/DataGrid/DataGrid.tsx
@@ -9,8 +9,8 @@ import type React from "react";
 import { useCallback, useEffect, useState } from "react";
 import _ from "underscore";
 
-import { useForceUpdate } from "metabase/common/hooks/use-force-update";
 import { getScrollBarSize } from "metabase/utils/dom";
+import { useForceUpdate } from "metabase/utils/use-force-update";
 
 import {
   ADD_COLUMN_BUTTON_WIDTH,
@@ -29,7 +29,7 @@ import type {
 import { AddColumnButton } from "../AddColumnButton/AddColumnButton";
 import { DataGridHeader } from "../DataGridHeader/DataGridHeader";
 import { DataGridRow } from "../DataGridRow/DataGridRow";
-import { Footer } from "../Footer/Footer";
+import { Footer, type FooterProps } from "../Footer/Footer";
 
 import S from "./DataGrid.module.css";
 import type { DataGridStylesProps } from "./types";
@@ -38,11 +38,12 @@ export interface DataGridProps<TData>
   extends DataGridInstance<TData>, DataGridStylesProps {
   emptyState?: React.ReactNode;
   showRowsCount?: boolean;
-  rowsTruncated?: number;
   isColumnReorderingDisabled?: boolean;
   theme?: DataGridTheme;
   zoomedRowIndex?: number;
   tableFooterExtraButtons?: React.ReactNode;
+  formatRowsCountMessage?: FooterProps<TData>["formatRowsCountMessage"];
+  formatPaginationMessage?: FooterProps<TData>["formatPaginationMessage"];
 }
 
 export const DataGrid = function DataGrid<TData>({
@@ -69,7 +70,8 @@ export const DataGrid = function DataGrid<TData>({
   onBodyCellClick,
   onWheel,
   tableFooterExtraButtons,
-  rowsTruncated,
+  formatRowsCountMessage,
+  formatPaginationMessage,
   onAddColumnClick,
   onHeaderCellClick,
   isColumnReorderingDisabled,
@@ -327,10 +329,11 @@ export const DataGrid = function DataGrid<TData>({
             table={table}
             enablePagination={enablePagination}
             showRowsCount={showRowsCount}
-            rowsTruncated={rowsTruncated}
             style={styles?.footer}
             className={classNames?.footer}
             tableFooterExtraButtons={tableFooterExtraButtons}
+            formatRowsCountMessage={formatRowsCountMessage}
+            formatPaginationMessage={formatPaginationMessage}
           />
         </div>
       </DndContext>
