+++ frontend/src/metabase/data-grid/index.ts
@@ -3,5 +3,6 @@ export * from "./components/DataGrid/types";
 export * from "./components/BaseCell/BaseCell";
 export * from "./components/HeaderCell/HeaderCell";
 export * from "./components/BodyCell/BodyCell";
+export * from "./components/PaginationFooter/PaginationFooter";
 export * from "./types";
 export * from "./hooks/use-data-grid-instance";
