+++ frontend/src/metabase/query_builder/containers/QueryBuilder.tsx
@@ -15,7 +15,6 @@ import { LeaveRouteConfirmModal } from "metabase/common/components/LeaveConfirmM
 import { isRouteInSync } from "metabase/common/hooks/is-route-in-sync";
 import { useCallbackEffect } from "metabase/common/hooks/use-callback-effect";
 import { useFavicon } from "metabase/common/hooks/use-favicon";
-import { useForceUpdate } from "metabase/common/hooks/use-force-update";
 import { useLoadingTimer } from "metabase/common/hooks/use-loading-timer";
 import { useWebNotification } from "metabase/common/hooks/use-web-notification";
 import { usePageTitleWithLoadingTime } from "metabase/hooks/use-page-title";
@@ -65,6 +64,7 @@ import {
   getUserIsAdmin,
 } from "metabase/selectors/user";
 import { getSetting } from "metabase/settings";
+import { useForceUpdate } from "metabase/utils/use-force-update";
 import type { Series } from "metabase-types/api";
 
 import {
