+++ frontend/src/metabase/utils/keyboard.ts
@@ -15,3 +15,12 @@ export const KEYCODE_FORWARD_SLASH = 191;
 export const KEY_ESCAPE = "Escape";
 export const KEY_ENTER = "Enter";
 export const KEY_BACKSPACE = "Backspace";
+
+/** Returns true if e.key is the given key and no modifier keys (ctrl, meta, alt, shift) were pressed */
+export const isPlainKey = (e: React.KeyboardEvent, key: string) => {
+  return e.key === key && !e.ctrlKey && !e.metaKey && !e.altKey && !e.shiftKey;
+};
+
+export const hasModifierKeys = (e: React.MouseEvent) => {
+  return e.shiftKey || e.ctrlKey || e.metaKey || e.altKey;
+};
