+++ frontend/src/metabase/utils/use-force-update.ts
