+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetail.tsx
@@ -1,12 +1,19 @@
 import { useEffect, useState } from "react";
+import { t } from "ttag";
 
+import { PaginationFooter } from "metabase/data-grid";
 import Question from "metabase-lib/v1/Question";
-
-import { PaginationFooter } from "../PaginationFooter/PaginationFooter";
+import { HARD_ROW_LIMIT } from "metabase-lib/v1/queries/utils";
 
 import S from "./ObjectDetail.module.css";
 import { ObjectDetailPanel } from "./ObjectDetailPanel";
 import type { ObjectDetailProps } from "./types";
+
+function getItemMessage(index: number, total: number) {
+  return total >= HARD_ROW_LIMIT
+    ? t`Item ${index + 1} of first ${total}`
+    : t`Item ${index + 1} of ${total}`;
+}
 
 export function ObjectDetail({
   question,
@@ -51,9 +58,9 @@ export function ObjectDetail({
           start={currentObjectIndex}
           end={currentObjectIndex}
           total={data.rows.length}
+          message={getItemMessage(currentObjectIndex, data.rows.length)}
           onNextPage={() => setCurrentObjectIndex((prev) => prev + 1)}
           onPreviousPage={() => setCurrentObjectIndex((prev) => prev - 1)}
-          singleItem
         />
       )}
     </>
